Move global vehicle enums to separate const.py
