+++ test/test_deprecated_vehicle.py
@@ -54,7 +54,7 @@ async def test_drive_train_attributes(caplog):
         VIN_G20: (True, False, False),
         VIN_G23: (False, True, False),
         VIN_I01_NOREX: (False, True, False),
-        VIN_I01_REX: (False, True, True),
+        VIN_I01_REX: (True, True, True),
         VIN_I20: (False, True, False),
     }
