+++ test/test_vehicle.py
@@ -3,7 +3,8 @@ import pytest
 
 from bimmer_connected.const import ATTR_ATTRIBUTES, ATTR_STATE, CarBrands
 from bimmer_connected.models import GPSPosition, StrEnum, VehicleDataBase
-from bimmer_connected.vehicle import DriveTrainType, VehicleViewDirection
+from bimmer_connected.vehicle import VehicleViewDirection
+from bimmer_connected.vehicle.const import DriveTrainType
 from bimmer_connected.vehicle.reports import CheckControlMessageReport
 
 from . import VIN_F31, VIN_G01, VIN_G20, VIN_G23, VIN_I01_NOREX, VIN_I01_REX, VIN_I20, get_deprecation_warning_count
@@ -77,7 +78,7 @@ async def test_drive_train_attributes(caplog):
         VIN_G20: (True, False, False),
         VIN_G23: (False, True, False),
         VIN_I01_NOREX: (False, True, False),
-        VIN_I01_REX: (False, True, True),
+        VIN_I01_REX: (True, True, True),
         VIN_I20: (False, True, False),
     }
